+++ app/src/main/java/com/ansgar/mylib/database/dao/CitationDao.java
@@ -20,6 +20,6 @@ public interface CitationDao {
 
     List<Citation> getAllCitations();
 
-    Observable<List<Citation>> getBookCitations(int bookId);
+    Observable<List<Citation>> getBookCitations(long bookId);
 
 }
