+++ app/src/main/java/com/ansgar/mylib/database/entity/Author.java
@@ -24,7 +24,7 @@ public class Author implements Serializable, Comparator<Author> {
 
     @SerializedName("id")
     @DatabaseField(generatedId = true, columnName = "id")
-    private int mId;
+    private long mId;
     @SerializedName("coverBytes")
     @DatabaseField(columnName = "cover_bytes")
     private String mCoverBytes;
@@ -55,7 +55,7 @@ public class Author implements Serializable, Comparator<Author> {
 
     }
 
-    public Author(int id, String firstName, String lastName, String biography, String date, User user, int hasSynchronized, List<Book> authorBooks) {
+    public Author(long id, String firstName, String lastName, String biography, String date, User user, int hasSynchronized, List<Book> authorBooks) {
         mId = id;
         mFirstName = firstName;
         mLastName = lastName;
@@ -66,11 +66,11 @@ public class Author implements Serializable, Comparator<Author> {
         mAuthorBooks = authorBooks;
     }
 
-    public int getId() {
+    public long getId() {
         return mId;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         mId = id;
     }
