+++ app/src/main/java/com/ansgar/mylib/ui/pager/BookCitationsPager.java
@@ -1,5 +1,13 @@
 package com.ansgar.mylib.ui.pager;
 
+import com.ansgar.mylib.R;
+import com.ansgar.mylib.ui.base.BaseFragment;
+import com.ansgar.mylib.ui.base.BasePresenter;
+import com.ansgar.mylib.ui.fragments.BookCitationsFragment;
+import com.ansgar.mylib.ui.fragments.BookDetailsFragment;
+import com.ansgar.mylib.ui.presenter.fragment.BookCitationsPagerPresenter;
+import com.ansgar.mylib.ui.view.fragment.BookCitationsPagerView;
+
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
@@ -8,16 +16,6 @@ import android.support.v4.view.ViewPager;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-
-import com.ansgar.mylib.R;
-import com.ansgar.mylib.database.entity.Book;
-import com.ansgar.mylib.ui.base.BaseFragment;
-import com.ansgar.mylib.ui.base.BasePresenter;
-import com.ansgar.mylib.ui.fragments.BookCitationsFragment;
-import com.ansgar.mylib.ui.fragments.BookDetailsFragment;
-import com.ansgar.mylib.ui.presenter.fragment.BookCitationsPagerPresenter;
-import com.ansgar.mylib.ui.view.fragment.BookCitationsPagerView;
-
 import butterknife.BindView;
 import butterknife.ButterKnife;
 
@@ -31,15 +29,15 @@ public class BookCitationsPager extends BaseFragment implements BookCitationsPag
     private static final String EXTRA_BOOK = "com.ansgar.mylib.ui.pager.book";
 
     private BookCitationsPagerPresenter mPresenter;
-    private int mBookId;
+    private long mBookId;
 
     @BindView(R.id.author_books_pager)
     ViewPager mAuthorBooksPager;
 
-    public static BookCitationsPager newInstance(int bookId) {
+    public static BookCitationsPager newInstance(long bookId) {
         BookCitationsPager pager = new BookCitationsPager();
         Bundle args = new Bundle();
-        args.putInt(EXTRA_BOOK, bookId);
+        args.putLong(EXTRA_BOOK, bookId);
         pager.setArguments(args);
         return pager;
     }
@@ -49,7 +47,7 @@ public class BookCitationsPager extends BaseFragment implements BookCitationsPag
     public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
         View view = inflater.inflate(LAYOUT, container, false);
         ButterKnife.bind(this, view);
-        mBookId = getArguments().getInt(EXTRA_BOOK);
+        mBookId = getArguments().getLong(EXTRA_BOOK);
         return view;
     }
